api_pipeline/requirements.txt
pyodbc
requests
memory-profiler
python-dotenv
